helpers/bq-remote-functions/get-buckets-metadata/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-storage
google-cloud-resource-manager
google-cloud-monitoring